codebase restructured, changed few file names (#14)
* moved utility `array` generators out of core/

* moved the code to different files

* restructured whole codebase
