+++ axon/csrc/core/core.cpp
@@ -1,10 +1,8 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
-#include <math.h>
 #include "core.h"
 #include "contiguous.h"
-#include "../cpu/helpers.h"
 
 Array* create_array(float* data, size_t ndim, int* shape, size_t size, dtype_t dtype) {
   if (data == NULL || !ndim || !size) {
@@ -570,113 +568,4 @@ void print_array(Array* self) {
   char result[8192] = "";
   format_array(self, self->shape, self->ndim, 0, 0, result);
   printf("axon.array(%s, dtype=%s)\n", result, get_dtype_name(self->dtype));
-}
-
-Array* zeros_like_array(Array* a) {
-  float* out = (float*)malloc(a->size * sizeof(float));
-  if (out == NULL) {
-    fprintf(stderr, "Memory allocation failed\n");
-    exit(EXIT_FAILURE);
-  }
-  zeros_like_array_ops(out, a->size);
-  Array* result = create_array(out, a->ndim, a->shape, a->size, a->dtype);
-  free(out);
-  return result;
-}
-
-Array* zeros_array(int* shape, size_t size, size_t ndim, dtype_t dtype) {
-  float* out = (float*)malloc(size * sizeof(float));
-  if (out == NULL) {
-    fprintf(stderr, "Memory allocation failed\n");
-    exit(EXIT_FAILURE);
-  }
-  zeros_array_ops(out, size);
-  Array* result = create_array(out, ndim, shape, size, dtype);
-  free(out);
-  return result;
-}
-
-Array* ones_like_array(Array* a) {
-  float* out = (float*)malloc(a->size * sizeof(float));
-  if (out == NULL) {
-    fprintf(stderr, "Memory allocation failed\n");
-    exit(EXIT_FAILURE);
-  }
-  ones_like_array_ops(out, a->size);
-  Array* result = create_array(out, a->ndim, a->shape, a->size, a->dtype);
-  free(out);
-  return result;
-}
-
-Array* ones_array(int* shape, size_t size, size_t ndim, dtype_t dtype) {
-  float* out = (float*)malloc(size * sizeof(float));
-  if (out == NULL) {
-    fprintf(stderr, "Memory allocation failed\n");
-    exit(EXIT_FAILURE);
-  }
-  ones_array_ops(out, size);
-  Array* result = create_array(out, ndim, shape, size, dtype);
-  free(out);
-  return result;
-}
-
-Array* randn_array(int* shape, size_t size, size_t ndim, dtype_t dtype) {
-  float* out = (float*)malloc(size * sizeof(float));
-  if (out == NULL) {
-    fprintf(stderr, "Memory allocation failed\n");
-    exit(EXIT_FAILURE);
-  }
-  fill_randn(out, size);
-  Array* result = create_array(out, ndim, shape, size, dtype);
-  free(out);
-  return result;
-}
-
-Array* randint_array(int low, int high, int* shape, size_t size, size_t ndim, dtype_t dtype) {
-  float* out = (float*)malloc(size * sizeof(float));
-  if (out == NULL) {
-    fprintf(stderr, "Memory allocation failed\n");
-    exit(EXIT_FAILURE);
-  }
-  fill_randint(out, low, high, size);
-  Array* result = create_array(out, ndim, shape, size, dtype);
-  free(out);
-  return result;
-}
-
-Array* uniform_array(int low, int high, int* shape, size_t size, size_t ndim, dtype_t dtype) {
-  float* out = (float*)malloc(size * sizeof(float));
-  if (out == NULL) {
-    fprintf(stderr, "Memory allocation failed\n");
-    exit(EXIT_FAILURE);
-  }
-  fill_uniform(out, low, high, size);
-  Array* result = create_array(out, ndim, shape, size, dtype);
-  free(out);
-  return result;
-}
-
-Array* fill_array(float fill_val, int* shape, size_t size, size_t ndim, dtype_t dtype) {
-  float* out = (float*)malloc(size * sizeof(float));
-  if (out == NULL) {
-    fprintf(stderr, "Memory allocation failed\n");
-    exit(EXIT_FAILURE);
-  }
-  fill_array_ops(out, fill_val, size);
-  Array* result = create_array(out, ndim, shape, size, dtype);
-  free(out);
-  return result;
-}
-
-Array* linspace_array(float start, float step, float end, int* shape, size_t size, size_t ndim, dtype_t dtype) {
-  float* out = (float*)malloc(size * sizeof(float));
-  if (out == NULL) {
-    fprintf(stderr, "Memory allocation failed\n");
-    exit(EXIT_FAILURE);
-  }
-  float step_size = (step > 1) ? (end - start) / (step - 1) : 0.0f;
-  linspace_array_ops(out, start, step_size, size);
-  Array* result = create_array(out, ndim, shape, size, dtype);
-  free(out);
-  return result;
 }
