+++ axon/csrc/cpu/ops_binary.cpp
@@ -1,7 +1,15 @@
 #include <stdio.h>
 #include <stddef.h>
 #include <math.h>
-#include "maths_ops.h"
+#include "ops_binary.h"
+
+void pow_array_ops(float* a, float exp, float* out, size_t size) {
+  for (size_t i = 0; i < size; i++) { out[i] = powf(a[i], exp); }
+}
+
+void pow_scalar_ops(float a, float* exp, float* out, size_t size) {
+  for (size_t i = 0; i < size; i++) { out[i] = powf(a, exp[i]); }
+}
 
 void add_ops(float* a, float* b, float* out, size_t size) {
   for (size_t i = 0; i < size; i++) { out[i] = a[i] + b[i]; }
@@ -59,50 +67,6 @@ void div_scalar_ops(float* a, float b, float* out, size_t size) {
       out[i] = a[i] / b;
     }
   }
-}
-
-void pow_array_ops(float* a, float exp, float* out, size_t size) {
-  for (size_t i = 0; i < size; i++) { out[i] = powf(a[i], exp); }
-}
-
-void pow_scalar_ops(float a, float* exp, float* out, size_t size) {
-  for (size_t i = 0; i < size; i++) { out[i] = powf(a, exp[i]); }
-}
-
-void exp_array_ops(float* a, float* out, size_t size) {
-  for (size_t i = 0; i < size; i++) { out[i] = expf(a[i]); }
-}
-
-void log_array_ops(float* a, float* out, size_t size) {
-  for (size_t i = 0; i < size; i++) { out[i] = logf(a[i]); }
-}
-
-void abs_array_ops(float* a, float* out, size_t size) {
-  for (size_t i = 0; i < size; i++) { out[i] = fabsf(a[i]); }
-}
-
-void sin_ops(float* a, float* out, size_t size) {
-  for (size_t i = 0; i < size; i++) { out[i] = sinf(a[i]); }
-}
-
-void cos_ops(float* a, float* out, size_t size) {
-  for (size_t i = 0; i < size; i++) { out[i] = cosf(a[i]); }
-}
-
-void tan_ops(float* a, float* out, size_t size) {
-  for (size_t i = 0; i < size; i++) { out[i] = tanf(a[i]); }
-}
-
-void sinh_ops(float* a, float* out, size_t size) {
-  for (size_t i = 0; i < size; i++) { out[i] = sinhf(a[i]); }
-}
-
-void cosh_ops(float* a, float* out, size_t size) {
-  for (size_t i = 0; i < size; i++) { out[i] = coshf(a[i]); }
-}
-
-void tanh_ops(float* a, float* out, size_t size) {
-  for (size_t i = 0; i < size; i++) { out[i] = tanhf(a[i]); }
 }
 
 void add_broadcasted_array_ops(float* a, float* b, float* out, int* broadcasted_shape, int broadcasted_size, int a_ndim, int b_ndim, int* a_shape, int* b_shape) {
