+++ axon/csrc/cpu/ops_binary.h
@@ -1,5 +1,5 @@
-#ifndef __MATHS_OPS__H__
-#define __MATHS_OPS__H__
+#ifndef __OPS_BINARY__H__
+#define __OPS_BINARY__H__
 
 #include <stddef.h>
 
@@ -14,16 +14,6 @@ extern "C" {
   void div_scalar_ops(float* a, float b, float* out, size_t size);
   void pow_array_ops(float* a, float exp, float* out, size_t size);
   void pow_scalar_ops(float a, float* exp, float* out, size_t size);
-  void exp_array_ops(float* a, float* out, size_t size);
-  void log_array_ops(float* a, float* out, size_t size);
-  void abs_array_ops(float* a, float* out, size_t size);
-
-  void sin_ops(float* a, float* out, size_t size);
-  void cos_ops(float* a, float* out, size_t size);
-  void tan_ops(float* a, float* out, size_t size);
-  void sinh_ops(float* a, float* out, size_t size);
-  void cosh_ops(float* a, float* out, size_t size);
-  void tanh_ops(float* a, float* out, size_t size);
 
   void add_broadcasted_array_ops(float* a, float* b, float* out, int* broadcasted_shape, int broadcasted_size, int a_ndim, int b_ndim, int* a_shape, int* b_shape);
   void sub_broadcasted_array_ops(float* a, float* b, float* out, int* broadcasted_shape, int broadcasted_size, int a_ndim, int b_ndim, int* a_shape, int* b_shape);
@@ -31,4 +21,4 @@ extern "C" {
   void mul_broadcasted_array_ops(float* a, float* b, float* out, int* broadcasted_shape, int broadcasted_size, int a_ndim, int b_ndim, int* a_shape, int* b_shape);
 }
 
-#endif  //!__OPS__H__
+#endif
