+++ axon/csrc/cpu/ops_unary.h
@@ -1,0 +1,19 @@
+#ifndef __OPS_UNARY__H__
+#define __OPS_UNARY__H__
+
+#include <stddef.h>
+
+extern "C" {
+  void exp_array_ops(float* a, float* out, size_t size);
+  void log_array_ops(float* a, float* out, size_t size);
+  void abs_array_ops(float* a, float* out, size_t size);
+
+  void sin_ops(float* a, float* out, size_t size);
+  void cos_ops(float* a, float* out, size_t size);
+  void tan_ops(float* a, float* out, size_t size);
+  void sinh_ops(float* a, float* out, size_t size);
+  void cosh_ops(float* a, float* out, size_t size);
+  void tanh_ops(float* a, float* out, size_t size);
+}
+
+#endif  //!__UNARY_OPS__H__
