+++ axon/csrc/unary_ops.cpp
@@ -1,0 +1,289 @@
+#include <stdio.h>
+#include <stdlib.h>
+#include "cpu/ops_unary.h"
+#include "unary_ops.h"
+
+Array* sin_array(Array* a) {
+  if (a == NULL) {
+    fprintf(stderr, "Array value pointers are null!\n");
+    exit(EXIT_FAILURE);
+  }
+  float* a_float = convert_to_float32(a->data, a->dtype, a->size);
+  if (a_float == NULL) {
+    fprintf(stderr, "Memory allocation failed during dtype conversion\n");
+    exit(EXIT_FAILURE);
+  }
+  float* out = (float*)malloc(a->size * sizeof(float));
+  if (out == NULL) {
+    fprintf(stderr, "Memory allocation failed\n");
+    free(a_float);
+    exit(EXIT_FAILURE);
+  }
+
+  sin_ops(a_float, out, a->size);
+  // For trigonometric functions, always return float type
+  // If input is integer, promote to float32; if already float, keep same precision
+  dtype_t result_dtype;
+  if (is_integer_dtype(a->dtype) || a->dtype == DTYPE_BOOL) {
+    result_dtype = DTYPE_FLOAT32;
+  } else {
+    result_dtype = a->dtype; // keep float32 or float64
+  }
+  Array* result = create_array(out, a->ndim, a->shape, a->size, result_dtype);
+  free(a_float);
+  free(out);
+  return result;
+}
+
+Array* sinh_array(Array* a) {
+  if (a == NULL) {
+    fprintf(stderr, "Array value pointers are null!\n");
+    exit(EXIT_FAILURE);
+  }
+  float* a_float = convert_to_float32(a->data, a->dtype, a->size);
+  if (a_float == NULL) {
+    fprintf(stderr, "Memory allocation failed during dtype conversion\n");
+    exit(EXIT_FAILURE);
+  }
+  float* out = (float*)malloc(a->size * sizeof(float));
+  if (out == NULL) {
+    fprintf(stderr, "Memory allocation failed\n");
+    free(a_float);
+    exit(EXIT_FAILURE);
+  }
+
+  sinh_ops(a_float, out, a->size);
+  // For trigonometric functions, always return float type
+  // If input is integer, promote to float32; if already float, keep same precision
+  dtype_t result_dtype;
+  if (is_integer_dtype(a->dtype) || a->dtype == DTYPE_BOOL) {
+    result_dtype = DTYPE_FLOAT32;
+  } else {
+    result_dtype = a->dtype; // keep float32 or float64
+  }
+  Array* result = create_array(out, a->ndim, a->shape, a->size, result_dtype);
+  free(a_float);
+  free(out);
+  return result;
+}
+
+Array* cos_array(Array* a) {
+  if (a == NULL) {
+    fprintf(stderr, "Array value pointers are null!\n");
+    exit(EXIT_FAILURE);
+  }
+  float* a_float = convert_to_float32(a->data, a->dtype, a->size);
+  if (a_float == NULL) {
+    fprintf(stderr, "Memory allocation failed during dtype conversion\n");
+    exit(EXIT_FAILURE);
+  }
+  float* out = (float*)malloc(a->size * sizeof(float));
+  if (out == NULL) {
+    fprintf(stderr, "Memory allocation failed\n");
+    free(a_float);
+    exit(EXIT_FAILURE);
+  }
+
+  cos_ops(a_float, out, a->size);
+  // For trigonometric functions, always return float type
+  // If input is integer, promote to float32; if already float, keep same precision
+  dtype_t result_dtype;
+  if (is_integer_dtype(a->dtype) || a->dtype == DTYPE_BOOL) {
+    result_dtype = DTYPE_FLOAT32;
+  } else {
+    result_dtype = a->dtype; // keep float32 or float64
+  }
+  Array* result = create_array(out, a->ndim, a->shape, a->size, result_dtype);
+  free(a_float);
+  free(out);
+  return result;
+}
+
+Array* cosh_array(Array* a) {
+  if (a == NULL) {
+    fprintf(stderr, "Array value pointers are null!\n");
+    exit(EXIT_FAILURE);
+  }
+  float* a_float = convert_to_float32(a->data, a->dtype, a->size);
+  if (a_float == NULL) {
+    fprintf(stderr, "Memory allocation failed during dtype conversion\n");
+    exit(EXIT_FAILURE);
+  }
+  float* out = (float*)malloc(a->size * sizeof(float));
+  if (out == NULL) {
+    fprintf(stderr, "Memory allocation failed\n");
+    free(a_float);
+    exit(EXIT_FAILURE);
+  }
+
+  cosh_ops(a_float, out, a->size);
+  // For trigonometric functions, always return float type
+  // If input is integer, promote to float32; if already float, keep same precision
+  dtype_t result_dtype;
+  if (is_integer_dtype(a->dtype) || a->dtype == DTYPE_BOOL) {
+    result_dtype = DTYPE_FLOAT32;
+  } else {
+    result_dtype = a->dtype; // keep float32 or float64
+  }
+  Array* result = create_array(out, a->ndim, a->shape, a->size, result_dtype);
+  free(a_float);
+  free(out);
+  return result;
+}
+
+Array* tan_array(Array* a) {
+  if (a == NULL) {
+    fprintf(stderr, "Array value pointers are null!\n");
+    exit(EXIT_FAILURE);
+  }
+  float* a_float = convert_to_float32(a->data, a->dtype, a->size);
+  if (a_float == NULL) {
+    fprintf(stderr, "Memory allocation failed during dtype conversion\n");
+    exit(EXIT_FAILURE);
+  }
+  float* out = (float*)malloc(a->size * sizeof(float));
+  if (out == NULL) {
+    fprintf(stderr, "Memory allocation failed\n");
+    free(a_float);
+    exit(EXIT_FAILURE);
+  }
+
+  tan_ops(a_float, out, a->size);
+  // For trigonometric functions, always return float type
+  // If input is integer, promote to float32; if already float, keep same precision
+  dtype_t result_dtype;
+  if (is_integer_dtype(a->dtype) || a->dtype == DTYPE_BOOL) {
+    result_dtype = DTYPE_FLOAT32;
+  } else {
+    result_dtype = a->dtype; // keep float32 or float64
+  }
+  Array* result = create_array(out, a->ndim, a->shape, a->size, result_dtype);
+  free(a_float);
+  free(out);
+  return result;
+}
+
+Array* tanh_array(Array* a) {
+  if (a == NULL) {
+    fprintf(stderr, "Array value pointers are null!\n");
+    exit(EXIT_FAILURE);
+  }
+  float* a_float = convert_to_float32(a->data, a->dtype, a->size);
+  if (a_float == NULL) {
+    fprintf(stderr, "Memory allocation failed during dtype conversion\n");
+    exit(EXIT_FAILURE);
+  }
+  float* out = (float*)malloc(a->size * sizeof(float));
+  if (out == NULL) {
+    fprintf(stderr, "Memory allocation failed\n");
+    free(a_float);
+    exit(EXIT_FAILURE);
+  }
+
+  tanh_ops(a_float, out, a->size);
+  // For trigonometric functions, always return float type
+  // If input is integer, promote to float32; if already float, keep same precision
+  dtype_t result_dtype;
+  if (is_integer_dtype(a->dtype) || a->dtype == DTYPE_BOOL) {
+    result_dtype = DTYPE_FLOAT32;
+  } else {
+    result_dtype = a->dtype; // keep float32 or float64
+  }
+  Array* result = create_array(out, a->ndim, a->shape, a->size, result_dtype);
+  free(a_float);
+  free(out);
+  return result;
+}
+
+Array* log_array(Array* a) {
+  if (a == NULL) {
+    fprintf(stderr, "Array value pointers are null!\n");
+    exit(EXIT_FAILURE);
+  }
+  float* a_float = convert_to_float32(a->data, a->dtype, a->size);
+  if (a_float == NULL) {
+    fprintf(stderr, "Memory allocation failed during dtype conversion\n");
+    exit(EXIT_FAILURE);
+  }
+  float* out = (float*)malloc(a->size * sizeof(float));
+  if (out == NULL) {
+    fprintf(stderr, "Memory allocation failed\n");
+    free(a_float);
+    exit(EXIT_FAILURE);
+  }
+
+  log_array_ops(a_float, out, a->size);
+  // If input is integer, promote to float32; if already float, keep same precision
+  dtype_t result_dtype;
+  if (is_integer_dtype(a->dtype) || a->dtype == DTYPE_BOOL) {
+    result_dtype = DTYPE_FLOAT32;
+  } else {
+    result_dtype = a->dtype; // keep float32 or float64
+  }
+  Array* result = create_array(out, a->ndim, a->shape, a->size, result_dtype);
+  free(a_float);
+  free(out);
+  return result;
+}
+
+Array* exp_array(Array* a) {
+  if (a == NULL) {
+    fprintf(stderr, "Array value pointers are null!\n");
+    exit(EXIT_FAILURE);
+  }
+  float* a_float = convert_to_float32(a->data, a->dtype, a->size);
+  if (a_float == NULL) {
+    fprintf(stderr, "Memory allocation failed during dtype conversion\n");
+    exit(EXIT_FAILURE);
+  }
+  float* out = (float*)malloc(a->size * sizeof(float));
+  if (out == NULL) {
+    fprintf(stderr, "Memory allocation failed\n");
+    free(a_float);
+    exit(EXIT_FAILURE);
+  }
+
+  exp_array_ops(a_float, out, a->size);
+  // If input is integer, promote to float32; if already float, keep same precision
+  dtype_t result_dtype;
+  if (is_integer_dtype(a->dtype) || a->dtype == DTYPE_BOOL) {
+    result_dtype = DTYPE_FLOAT32;
+  } else {
+    result_dtype = a->dtype; // keep float32 or float64
+  }
+  Array* result = create_array(out, a->ndim, a->shape, a->size, result_dtype);
+  free(a_float);
+  free(out);
+  return result;
+}
+
+Array* abs_array(Array* a) {
+  if (a == NULL) {
+    fprintf(stderr, "Array value pointers are null!\n");
+    exit(EXIT_FAILURE);
+  }
+  float* a_float = convert_to_float32(a->data, a->dtype, a->size);
+  if (a_float == NULL) {
+    fprintf(stderr, "Memory allocation failed during dtype conversion\n");
+    exit(EXIT_FAILURE);
+  }
+  float* out = (float*)malloc(a->size * sizeof(float));
+  if (out == NULL) {
+    fprintf(stderr, "Memory allocation failed\n");
+    free(a_float);
+    exit(EXIT_FAILURE);
+  }
+
+  abs_array_ops(a_float, out, a->size);
+  // If input is integer, promote to float32; if already float, keep same precision
+  dtype_t result_dtype;
+  if (is_integer_dtype(a->dtype) || a->dtype == DTYPE_BOOL) {
+    result_dtype = DTYPE_FLOAT32;
+  } else {
+    result_dtype = a->dtype; // keep float32 or float64
+  }
+  Array* result = create_array(out, a->ndim, a->shape, a->size, result_dtype);
+  free(a_float);
+  free(out);
+  return result;
+}
